Refactored code to isolate models and repository. Repository is now an interface and I’ve provided an implementation for it.
